feat(manage): show progress in 'update all' button (#1333)
* move snap l10n extensions into separate file

* show change progress in 'update all' button

* update tests, leave todo notes

* add 'updating' l10n string

* re-use updatesModel
