+++ lib/src/detail/detail_page.dart
@@ -509,20 +509,3 @@ class _ChannelDropdownEntry extends StatelessWidget {
     );
   }
 }
-
-extension SnapdChangeL10n on SnapdChange {
-  String? localize(AppLocalizations l10n) => switch (kind) {
-        'install-snap' => l10n.snapActionInstallingLabel,
-        'remove-snap' => l10n.snapActionRemovingLabel,
-        _ => null,
-      };
-}
-
-extension SnapConfinementL10n on SnapConfinement {
-  String localize(AppLocalizations l10n) => switch (this) {
-        SnapConfinement.classic => l10n.snapConfinementClassic,
-        SnapConfinement.devmode => l10n.snapConfinementDevmode,
-        SnapConfinement.strict => l10n.snapConfinementStrict,
-        _ => name,
-      };
-}
